Ending of refactoring

+++ app/Repositories/RipRepository.php
@@ -6,24 +6,34 @@ use App\Repositories\Interfaces\RipRepositoryInterface;
 use Carbon\Carbon;
 use App\Rip;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Log;
 
 class RipRepository implements RipRepositoryInterface
 {
 
-    public function store(int $id): void
+    public function store(int $userId): void
     {
         $rip = new Rip;
-        $rip->user_id = $id;
+        $rip->user_id = $userId;
         $rip->rip_date = Carbon::now();
         $rip->save();
     }
 
-    public function delete(int $id): Model
+    public function delete(int $userId): void
     {
-        $rip = Rip::where('user_id', $id)->firstOrFail();
+        $rip = Rip::where('user_id', $userId)->firstOrFail();
 
-        $rip->delete();
+        if ($rip instanceof Model) {
+            try {
+                $rip->delete();
+            } catch (\Exception $e) {
+                Log::error($e->getMessage());
+            }
+        }
+    }
 
-        return $rip;
+    public function getAll(): array
+    {
+        return Rip::all()->toArray();
     }
 }
